+++ GrowthBook/ExperimentResult.cs
@@ -54,7 +54,7 @@ namespace GrowthBook
         /// <summary>
         /// The hash value used to assign a variation (float from 0 to 1).
         /// </summary>
-        public float Bucket { get; set; }
+        public double Bucket { get; set; }
 
         /// <summary>
         /// The human-readable name of the assigned variation.
