+++ GrowthBook/FeatureRule.cs
@@ -20,7 +20,7 @@ namespace GrowthBook
         /// <summary>
         /// What percent of users should be included in the experiment (between 0 and 1, inclusive).
         /// </summary>
-        public float? Coverage { get; set; }
+        public double? Coverage { get; set; }
 
         /// <summary>
         /// Immediately force a specific value (ignore every other option besides condition and coverage).
@@ -40,7 +40,7 @@ namespace GrowthBook
         /// <summary>
         /// How to weight traffic between variations. Must add to 1.
         /// </summary>
-        public List<float> Weights { get; set; }
+        public List<double> Weights { get; set; }
 
         /// <summary>
         /// Adds the experiment to a namespace.
